+++ newton/usd.py
@@ -58,7 +58,6 @@ __all__ = [
 from ._src.usd.schema_resolver import (
     PrimType,
     SchemaResolver,
-    create_schema_resolution,
 )
 from ._src.usd.schemas import (
     SchemaResolverMjc,
@@ -72,5 +71,4 @@ __all__ += [
     "SchemaResolverMjc",
     "SchemaResolverNewton",
     "SchemaResolverPhysx",
-    "create_schema_resolution",
 ]